correção de layout e adição de cabeçalho

+++ Front-end/js/dados.js
@@ -1,6 +1,19 @@
+document.getElementById("linkVoltar").addEventListener("click", function () {
+    window.location.href = "perfil.html";
+});
+
+document.querySelectorAll("#marca, #linkLogo").forEach(element => {
+    element.addEventListener("click", function () {
+        window.location.href = "solicitacao.html";
+    });
+});
+
 $(document).ready(function () {
     const token = localStorage.getItem('token');
     const userId = localStorage.getItem('userId');
+    const errorDelete = document.getElementById("error-message-delete");
+
+    errorDelete.textContent = "";
 
     if (token && userId) {
         const url = `http://localhost:3000/buscar/conta/${userId}`;
@@ -24,40 +37,49 @@ $(document).ready(function () {
     }
 
     // deletar conta
-    $("#deletar_conta").click(function() {
-        if (confirm("Você tem certeza que deseja deletar sua conta? Esta ação não pode ser desfeita.")) {
-            if (token) {
-                $.ajax({
-                    url: `http://localhost:3000/deletar/conta/${userId}`,
-                    type: 'DELETE',
-                    headers: {
-                        'Authorization': 'Bearer ' + token
-                    },
-                    success: function() {
-                        $.ajax({
-                            url: 'http://localhost:3000/logout',
-                            type: 'POST',
-                            headers: {
-                                'Authorization': 'Bearer ' + token
-                            },
-                            success: function() {
-                                localStorage.removeItem('token');
-                                localStorage.removeItem('userId');
-                                alert('Conta deletada com sucesso.');
-                                window.location.href = 'login.html';
-                            },
-                            error: function(jqXHR, textStatus, errorThrown) {
-                                console.error("Erro ao fazer logout:", errorThrown);
-                            }
-                        });
-                    },
-                    error: function(jqXHR, textStatus, errorThrown) {
-                        console.error("Erro ao deletar conta:", errorThrown);
-                    }
-                });
-            } else {
-                console.error("Token não encontrado no localStorage.");
-            }
+    $("#deletar_conta").click(function () {
+        $('#confirmDeleteModal').modal('show');
+    });
+
+    $("#confirmDeleteButton").click(function () {
+        var token = localStorage.getItem('token');
+        var userId = localStorage.getItem('userId');
+
+        if (token && userId) {
+            $.ajax({
+                url: `http://localhost:3000/deletar/conta/${userId}`,
+                type: 'DELETE',
+                headers: {
+                    'Authorization': 'Bearer ' + token
+                },
+                success: function () {
+                    $.ajax({
+                        url: 'http://localhost:3000/logout',
+                        type: 'POST',
+                        headers: {
+                            'Authorization': 'Bearer ' + token
+                        },
+                        success: function () {
+                            localStorage.removeItem('token');
+                            localStorage.removeItem('userId');
+                            $('#error-message-delete').text("Conta deletada com sucesso");
+                            setTimeout(function () {
+                                window.location.href = "login.html";
+                            }, 1000);
+                        },
+                        error: function (jqXHR, textStatus, errorThrown) {
+                            console.error("Erro ao fazer logout:", errorThrown);
+                        }
+                    });
+                },
+                error: function (jqXHR, textStatus, errorThrown) {
+                    console.error("Erro ao deletar conta:", errorThrown);
+                }
+            });
+        } else {
+            console.error("Token não encontrado no localStorage.");
         }
+
+        $('#confirmDeleteModal').modal('hide');
     });
 });
